+++ packages/components/atoms/f-popover/src/components/_tests/Popover.test.js
@@ -1,0 +1,10 @@
+import { shallowMount } from '@vue/test-utils';
+import VPopover from '../Popover.vue';
+
+describe('Popover', () => {
+    it('should be defined', () => {
+        const propsData = {};
+        const wrapper = shallowMount(VPopover, { propsData });
+        expect(wrapper.exists()).toBe(true);
+    });
+});