+++ packages/components/organisms/f-checkout/src/components/Checkout.vue
@@ -193,6 +193,16 @@ export default {
         applicationName: {
             type: String,
             required: true
+        },
+
+        getGeoLocationUrl: {
+            type: String,
+            required: true
+        },
+
+        getGeoLocationTimeout: {
+            type: Number,
+            default: 1000
         }
     },
 
@@ -240,7 +250,8 @@ export default {
             'time',
             'userNote',
             'basket',
-            'orderId'
+            'orderId',
+            'geolocation'
         ]),
 
         isMobileNumberValid () {
@@ -298,7 +309,8 @@ export default {
             'setAuthToken',
             'updateCustomerDetails',
             'updateUserNote',
-            'placeOrder'
+            'placeOrder',
+            'getGeoLocation'
         ]),
 
         ...mapActions('analytics', [
@@ -343,12 +355,15 @@ export default {
                     await this.setupGuestUser();
                 }
 
+                await this.lookupGeoLocation();
+
                 const data = mapUpdateCheckoutRequest({
                     address: this.address,
                     customer: this.customer,
                     isCheckoutMethodDelivery: this.isCheckoutMethodDelivery,
                     time: this.time,
-                    userNote: this.userNote
+                    userNote: this.userNote,
+                    geolocation: this.geolocation
                 });
 
                 await this.updateCheckout({
@@ -532,11 +547,37 @@ export default {
                     language: this.$i18n.locale,
                     timeout: this.checkoutTimeout
                 });
-
                 this.$emit(EventNames.CheckoutAddressGetSuccess);
             } catch (thrownErrors) {
                 this.$emit(EventNames.CheckoutAddressGetFailure, thrownErrors);
                 this.hasCheckoutLoadedSuccessfully = false;
+            }
+        },
+
+        /**
+         * Look up the geo details for the customers address, skip on failure.
+         *
+         */
+        async lookupGeoLocation () {
+            const locationData =
+            {
+                addressLines: Object.values(this.address).filter(addressLine => !!addressLine)
+            };
+
+            try {
+                if (locationData.addressLines.length) {
+                    await this.getGeoLocation({
+                        url: this.getGeoLocationUrl,
+                        postData: locationData,
+                        timeout: this.getGeoLocationTimeout
+                    });
+                }
+            } catch (error) {
+                this.$logger.logWarn(
+                    'Geo location lookup failed',
+                    this.$store,
+                    { error }
+                );
             }
         },
 
@@ -714,7 +755,7 @@ export default {
     .c-spinner {
         margin: 0 auto;
     }
-}
+  }
 
 .c-checkout {
     padding-top: spacing(x6);